Introduced styling via Styled Components
Currently this throws a console error:

Warning: React attempted to reuse markup in a container but the checksum was invalid. This generally means that you are using server rendering and the markup generated on the server was not what the client was expecting. React injected new markup to compensate which works but you have lost many of the benefits of server rendering. Instead, figure out why the markup being generated is different on the client or server:

(client) ="20"><li class="sc-bdVaJa dfQcku" data-
(server) ="20"><li class="sc-htpNat ghdkrB" data-