WIP
This is a messy WIP branch to document the pairing session @starsirius
and I had in order to investigate https://github.com/artsy/force/issues/17

So far our lead is that the list of cities — which we fetch from S3,
and use to map city slugs to coordinates — is not in scope early enough.

A server side render of a /galleries page where there is a "location"
param in the query string would have to translate that "location" slug
into coordinates that can be passed to Gravity's /v1/partners endpoint.

We started to rewrite the route handler to fetch the cities before
proceeding with the rest of the query.