Pass hubs data without using Sharify
My understanding is that Sharify is only needed when data is to be
shared between a server-rendered view and a client-rendered view.

In the current case we have only a server-rendered Jade template, so the
data can be passed simply as a local variable to the render method.

We are not shipping any client-side code for the homepage hubs nav
(right?) and therefore there is no hydration step. Thus no need to
access this data on the client side.